requirements.txt
langchain
llama-index
chromadb
pydantic
python-dotenv
flask
streamlit
ollama
pywhisper
beautifulsoup4
markdown
numpy
pandas
python-frontmatter
tqdm
flask-cors
frontmatter
beautifulsoup4
